backend/data-utils/requirements.txt
pymongo[srv]
qdrant-client
fastembed
python-dotenv
bcrypt